requirements.txt
protobuf==3.20.2
google-ads
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
python-dateutil
pyyaml
